added null checks to write/destroy jsprincipal functions

+++ src/jsglue.cpp
@@ -353,11 +353,14 @@ class ServoJSPrincipal : public JSPrincipals
     }
 
     virtual void destroy() {
-      this->destroyCallback(this);
+      if(this->destroyCallback)
+        this->destroyCallback(this);
     }
 
     bool write(JSContext* cx, JSStructuredCloneWriter* writer) {
-      return this->writeCallback(cx, writer);
+      return this->writeCallback
+             ? this->writeCallback(cx, writer)
+             : false;
     }
 };
